requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
mysql-connector-python==9.2.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
pygame==2.6.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
